# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Elite FF Tournament Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments:
    get:
      operationId: getTournaments
      tags: [tournaments]
      summary: List all tournaments
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tournaments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tournament"
    post:
      operationId: createTournament
      tags: [tournaments]
      summary: Create a tournament (Host only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTournamentBody"
      responses:
        "201":
          description: Tournament created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
  /tournaments/{id}:
    get:
      operationId: getTournament
      tags: [tournaments]
      summary: Get tournament details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "404":
          description: Not found
    put:
      operationId: updateTournament
      tags: [tournaments]
      summary: Update a tournament (Host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTournamentBody"
      responses:
        "200":
          description: Tournament updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
    delete:
      operationId: deleteTournament
      tags: [tournaments]
      summary: Delete a tournament (Host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tournaments/{id}/cancel:
    post:
      operationId: cancelTournament
      tags: [tournaments]
      summary: Cancel a tournament (Host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelTournamentBody"
      responses:
        "200":
          description: Tournament cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
  /tournaments/{id}/delay:
    post:
      operationId: delayTournament
      tags: [tournaments]
      summary: Delay a tournament (Host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DelayTournamentBody"
      responses:
        "200":
          description: Tournament delayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
  /tournaments/{id}/complete:
    post:
      operationId: completeTournament
      tags: [tournaments]
      summary: Mark tournament as completed (Host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
  /tournaments/{id}/room:
    post:
      operationId: uploadRoomCredentials
      tags: [tournaments]
      summary: Upload room ID and password (Host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomCredentialsBody"
      responses:
        "200":
          description: Room credentials uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
  /tournaments/{id}/scoreboard:
    get:
      operationId: getScoreboard
      tags: [scoreboard]
      summary: Get live scoreboard for a tournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scoreboard data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreboardEntry"
    post:
      operationId: updateScoreboard
      tags: [scoreboard]
      summary: Update scoreboard entry (Host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreboardUpdateBody"
      responses:
        "200":
          description: Scoreboard updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreboardEntry"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Tournament:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        mode:
          type: string
        mapName:
          type: ["string", "null"]
        matchCount:
          type: integer
        maps:
          type: ["string", "null"]
        killPoints:
          type: integer
        placements:
          type: ["string", "null"]
        prizeDistribution:
          type: ["string", "null"]
        teamSize:
          type: string
        entryFee:
          type: ["integer", "null"]
        prizePool:
          type: ["integer", "null"]
        booyahPrize:
          type: ["integer", "null"]
        secondPrize:
          type: ["integer", "null"]
        thirdPrize:
          type: ["integer", "null"]
        highestKillPrize:
          type: ["integer", "null"]
        maxSlots:
          type: integer
        filledSlots:
          type: integer
        status:
          type: string
        scheduledAt:
          type: string
        rules:
          type: ["string", "null"]
        posterUrl:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        qrUrl:
          type: ["string", "null"]
        roomId:
          type: ["string", "null"]
        roomPassword:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        delayInfo:
          type: ["string", "null"]
        isPaid:
          type: boolean
        timerEnabled:
          type: boolean
        isPrivate:
          type: boolean
        inviteLink:
          type: ["string", "null"]
        createdAt:
          type: string
        hostId:
          type: integer
      required:
        - id
        - name
        - type
        - mode
        - teamSize
        - maxSlots
        - filledSlots
        - status
        - scheduledAt
        - isPaid
        - timerEnabled
        - isPrivate
        - createdAt
        - hostId
    CreateTournamentBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        mode:
          type: string
        mapName:
          type: ["string", "null"]
        matchCount:
          type: integer
        maps:
          type: ["array", "null"]
          items:
            type: string
        killPoints:
          type: integer
        placements:
          type: ["array", "null"]
          items:
            type: integer
        prizeDistribution:
          type: ["array", "null"]
          items:
            type: object
            properties:
              rank:
                type: integer
              label:
                type: string
              amount:
                type: integer
            required:
              - rank
              - label
              - amount
        teamSize:
          type: string
        entryFee:
          type: ["integer", "null"]
        prizePool:
          type: ["integer", "null"]
        booyahPrize:
          type: ["integer", "null"]
        secondPrize:
          type: ["integer", "null"]
        thirdPrize:
          type: ["integer", "null"]
        highestKillPrize:
          type: ["integer", "null"]
        maxSlots:
          type: integer
        scheduledAt:
          type: string
        rules:
          type: ["string", "null"]
        posterUrl:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        isPaid:
          type: boolean
        timerEnabled:
          type: boolean
        isPrivate:
          type: boolean
      required:
        - name
        - type
        - mode
        - teamSize
        - maxSlots
        - scheduledAt
        - isPaid
        - timerEnabled
    UpdateTournamentBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        mode:
          type: ["string", "null"]
        mapName:
          type: ["string", "null"]
        teamSize:
          type: ["string", "null"]
        entryFee:
          type: ["integer", "null"]
        prizePool:
          type: ["integer", "null"]
        booyahPrize:
          type: ["integer", "null"]
        secondPrize:
          type: ["integer", "null"]
        thirdPrize:
          type: ["integer", "null"]
        highestKillPrize:
          type: ["integer", "null"]
        maxSlots:
          type: ["integer", "null"]
        scheduledAt:
          type: ["string", "null"]
        rules:
          type: ["string", "null"]
        posterUrl:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        isPaid:
          type: ["boolean", "null"]
        timerEnabled:
          type: ["boolean", "null"]
        roomId:
          type: ["string", "null"]
        roomPassword:
          type: ["string", "null"]
    CancelTournamentBody:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
    DelayTournamentBody:
      type: object
      properties:
        newScheduledAt:
          type: string
        reason:
          type: string
      required:
        - newScheduledAt
        - reason
    RoomCredentialsBody:
      type: object
      properties:
        roomId:
          type: string
        roomPassword:
          type: string
      required:
        - roomId
        - roomPassword
...
    ScoreboardEntry:
      type: object
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        registrationId:
          type: integer
        squadName:
          type: string
        kills:
          type: integer
        rank:
          type: ["integer", "null"]
        points:
          type: integer
        updatedAt:
          type: string
      required:
        - id
        - tournamentId
        - registrationId
        - squadName
        - kills
        - points
        - updatedAt
    ScoreboardUpdateBody:
      type: object
      properties:
        registrationId:
          type: integer
        squadName:
          type: string
        kills:
          type: integer
        rank:
          type: ["integer", "null"]
        points:
          type: integer
      required:
        - registrationId
        - squadName
        - kills
        - points